[REFACTOR][#72] Filter - inputValue custom hook에 요청을 보내는 함수 추가
- inputValue custom hook에 요청을 보내는 함수 추가
   기존 inputBox에 있던 `synchronizeModel`, `keyPress`, 함수를 custom hook으로 이동
   custom hook에서 issue, filter model 읽어옴
   event 객체를 받아 (input 후 엔터) 요청을 보내는 함수와
   input의 값을 받아 요청을 보내는 함수 생성

- Dropdown 버튼이 눌릴 시, Filter에서 Dropdown 버튼의 함수와, inputValue 함수를 합성해서 요청을 보내도록 기능 구현
